Move EditClientCommand to client package
